Update Redis command handlers for better performance and reliability, Add EXEC command
