+++ app/transactions/exec-handler.go
@@ -1,0 +1,7 @@
+package transactions
+
+import "github.com/jgrecu/redis-clone/app/resp"
+
+func Exec(params []resp.RESP) []byte {
+    return resp.Error("ERR EXEC without MULTI").Marshal()
+}
